move memory mgmt syscalls to litebox_common_linux
